+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 0.8.0 (September 30, 2025)
+- Add 49 new variables (from WSC team) and data dictionary entries
+- Add questionnaire dataset and data dictionary entries
+- The CSV datasets generated from a new R script is located here:
+  - `\\rfawin\BWH-SLEEPEPI-NSRR-STAGING\20200115-peppard-wsc\nsrr-prep\_releases\0.8.0.pre`
+
+
 ## 0.7.0 (July 17, 2024)
 - Remove death data (WSC contributor request)
 - Add ahi3 variable